feat: Add Windows ARM64 (Snapdragon X) support
This commit adds full support for building and running DiskANN on Windows ARM64
platforms, specifically targeting Qualcomm Snapdragon X processors.

Key changes:
- Add build.bat and test.bat scripts for Windows ARM64 compilation
  - Properly configure Visual Studio 2022 ARM64 environment
  - Set Windows SDK library paths to resolve linking issues
  - Work around Git bash link.exe conflict with MSVC linker

- Fix compilation errors for ARM64 Windows target:
  - Fix type inference issues in graph module (vamana.rs, prune.rs)
  - Fix Send/Sync trait implementation for Qualcomm DirectML types
  - Add missing create_distance_function export in lib.rs
  - Fix test compilation errors in labels and integration tests

- Update CLI binary to avoid duplicate Cli struct definition
- Remove unused imports and variables to clean up warnings

The project now successfully builds and all tests pass on Windows ARM64.

+++ tests/integration_tests.rs
@@ -156,7 +156,7 @@ fn test_large_scale_index() {
         total_search_time += start.elapsed();
     }
     
-    let avg_search_time = total_search_time / num_queries;
+    let avg_search_time = total_search_time / num_queries as u32;
     let qps = 1_000_000.0 / avg_search_time.as_micros() as f64;
     
     println!("Average search time: {:?}", avg_search_time);
